Reorganize Menu for better UX
Moved around a few things and added back the home button to make more space, and to give a more natural UX than what it currently is. The "N.T.O.S. - TMU Notices" is barely used, especially by pilots - and takes up twice the space that of any other menu item, so it's been moved to a sub-menu. It's also not important to have that button there, since any relevant TMU notices gets shown on the front page anyway, so 90% of the week, it is a useless button.